fix: :lock: Make sure to use a SecureString when passing around the password
Using a SecureString we ensure that the password will not be left over in memory at a later point. While this is not 100% secure it should work in the majority of cases. However as we talk to postgres there ma be some leakage. However at that point we already have a hash which is compared. Not the password itself.